combine separate class file.
